+++ conans/test/integration/conanfile/set_name_version_test.py
@@ -150,29 +150,3 @@ class SetVersionNameTest(unittest.TestCase):
             client.save({"version.txt": "2.1"}, clean_first=True)
             client.run("export .. ")
             self.assertIn("pkg/2.1: A new conanfile.py version was exported", client.out)
-
-    def test_capture_git_tag(self):
-        """
-        A local repo, without remote, will have commit, but no URL
-        """
-        c = TestClient()
-        conanfile = textwrap.dedent("""
-            from conan import ConanFile
-            from conan.tools.scm import Git
-
-            class Pkg(ConanFile):
-                name = "pkg"
-
-                def set_version(self):
-                    git = Git(self, self.recipe_folder)
-                    self.version = git.run("describe --tags")
-            """)
-        c.save({"conanfile.py": conanfile})
-        c.init_git_repo()
-        c.run_command("git tag 1.2")
-        c.run("install .")
-        assert "pkg/1.2" in c.out
-        c.run("create .")
-        assert "pkg/1.2" in c.out
-        c.run("install --requires=pkg/1.2")
-        assert "pkg/1.2" in c.out
